Ensure direnv runs before everything else
If you're trying to load dev tools from your direnv, this plugin needs
to run first and block while it's running the command. Loading on
filesystem and using execSync makes sure it runs before everything else.

Also, update the typescript/vscode version because I was getting errors
and that somehow fixed it.